+++ kymatio/scattering1d/frontend/numpy_frontend.py
@@ -75,7 +75,7 @@ class TimeFrequencyScatteringNumPy1D(TimeFrequencyScatteringBase1D,
                  average=True, average_fr=False, oversampling=0,
                  oversampling_fr=None, aligned=True,
                  sampling_filters_fr='resample', out_type="array", out_3D=False,
-                 out_exclude=None, pad_mode='zero',
+                 out_exclude=None, pad_mode='reflect',
                  pad_mode_fr='conj-reflect-zero', max_pad_factor=2,
                  max_pad_factor_fr=None, backend="numpy"):
         if oversampling_fr is None:
