+++ kymatio/scattering1d/frontend/tensorflow_frontend.py
@@ -77,7 +77,7 @@ class TimeFrequencyScatteringTensorFlow1D(TimeFrequencyScatteringBase1D,
                  average=True, average_fr=False, oversampling=0,
                  oversampling_fr=None, aligned=True, sampling_filters_fr='resample',
                  out_type="array", out_3D=False, out_exclude=None,
-                 pad_mode='zero', max_pad_factor=2, max_pad_factor_fr=None,
+                 pad_mode='reflect', max_pad_factor=2, max_pad_factor_fr=None,
                  pad_mode_fr='conj-reflect-zero', backend='tensorflow',
                  name='TimeFrequencyScattering1D'):
         if oversampling_fr is None:
